Fix a threading issue for customized audio device module.
Customized audio device module uses a frame generator as input and
a default internal ADM as output. However, default ADM only allows to be
called on the same thread as it is created. This change creates default
internal ADM when it is first used instead of when the customized ADM is
created.

Change-Id: I933d395bb4c86b8ba44127a2cf653e4cf0f92ba1
Reviewed-on: https://git-ccr-1.devtools.intel.com/gerrit/33915
Tested-by: webrtctest <webrtctest@intel.com>
Reviewed-by: Zhang, Wenjie <wenjie.zhang@intel.com>
Tested-by: Zhang, Wenjie <wenjie.zhang@intel.com>