Spliting SDK and WebRTC.
Before this change, WooGeen SDK is mixed into WebRTC's codebase, so we can
utilize WebRTC's toolchain. From now on, we'll place WebRTC as a third party
dependency. But we still use depot_tools and GN for code management and
building.

This change adds DEPS file which includes libraies that WooGeen SDK depends on,
as well as GN files for building WebRTC and WooGeen SDK.

Change-Id: Id4307efe2c856262d86388775afe7373e5cedbc7
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/83910
Reviewed-by: Qiu, Jianlin <jianlin.qiu@intel.com>
Tested-by: Qiu, Jianlin <jianlin.qiu@intel.com>