Before exiting the CustomizedFramesThread, call the Cleanup() (#400)
* Before exiting the CustomizedFramesThread, call the Cleanup()
method on the FrameGeneratorInterface. Needed for generators that
use thread-senstive resources.

To ensure resources are constructed/destructed on a single thread,
use a bool that is set on the first call to GenerateNextFrame to
construct, and call destructs in the Cleanup() method. Cleanup() has
a default noop implementation for backwards-compatibility.

* Remove bad ; from interface method.