Added public method for changing ICE servers on existing P2PClient
Base OWT framework had an issue where closing a connection would deadlock,
because the webrtc signaling_thread was waiting on itself (you can't call Close()
on the signaling thread, it's doing the work). They did a bad fix of intentionally
'leaking' the removed P2PPeerConnectionChannels by dumping them to a vector.
I fixed it by instead removing the channel in a detached thread. The OnStopped()
also calls the SignalingChannel's SendMessage, which is intercepted, to properly
release resources on streams when a connection disconnects without warning.

I plan to submit the detached erase as another PR to base OWT.