Stable, synchronous closing of peer connections.
Fixing the deadlock in OnStopped revealed an underlying bug in OWT that was segfaulting whenever a connection was closed.
The OnStopped() method deletes the P2PPeerConnectionChannel, which is the PeerConnectionObserver. OnStopped() was called
by TriggerOnStopped(), which was called in the OnIceConnectionChange(kClosed) function. This meant that the Observer was
being deleted during the execution of peer_connection->Close(), instead of waiting until the Close() returned, when it's
safe to delete the observer. This means removing TriggerOnStopped() from the OnIceConnectionChange callback.

Additionally, the previous hack to "intercept" outgoing messages in OnStopped() is totally unneeded, as the SendStop() in
P2PPeerConnection::Stop() already sends such a message in the framework. This message will be the one intercepted, which
is when the AmbientPublication references to the Frame Generators are deleted.

The final change is reversing the ordering of SendStop() and ClosePeerConnection() in Stop(). In order for the SHM readers to
be released on the correct thread, the final reference to the streams must be inside the peer connection, not the AmbientPublication.
Therefore, SendStop() (deleting the Publication) is moved before ClosePeerConnection() (closing the peer connection, releasing the SHM
readers)