Previous change to ClosePeerConnection was correct. TriggerOnStopped() must only be called after peer_connection_->Close() returns, not in OnIceConnectionChange(), which is called inside Close(). Instead, ended_ flag is set in OnIceConnectionChange(), so as soon as Close() starts, nothing more operations on the connection can be performed (Close() still needs to return before TriggerOnStopped() or the P2PPeerConnectionChannel is destroyed).
Previously, these changes failed because P2PClient::OnStopped() would immediately delete
the last reference to the pointer, which would delete the object while it was still on
the stack of functions. But, now OnStopped() doesn't delete the stopped channel until the
subsequent channel stops. Because the connection closes, streams and SHM readers are not
leaked between calls to OnStopped().