logging: use static stream in LogToFileRotate (#718)
We initialized the stream to write to in LogToFileRotate() but
it got deinitialized on the exit from the scope while its
reference remained in the logging which causes a crash on the
first write. That's a quick fix. More appropriately would be
to change webrtc function prototype for it to accept stream
for the ref count increment.

Signed-off-by: Dmitry Rogozhkin <dmitry.v.rogozhkin@intel.com>